Update the error codes & introduce role entity.
