+++ components/workflow-mgt/org.wso2.carbon.identity.workflow.mgt/src/main/java/org/wso2/carbon/identity/workflow/mgt/bean/Entity.java
@@ -26,9 +26,6 @@ public class Entity {
     private String entityId;
     private String entityType;
     private int tenantId;
-    private String tenantDomain;
-    private String audience;
-    private String audienceId;
 
     public Entity(String entityId, String entityType, int tenantId) {
 
@@ -37,21 +34,10 @@ public class Entity {
         this.tenantId = tenantId;
     }
 
-    public Entity(String entityId, String entityType, int tenantId, String tenantDomain, String audience,
-                  String audienceId) {
-
-        this.entityId = entityId;
-        this.entityType = entityType;
-        this.tenantId = tenantId;
-        this.tenantDomain = tenantDomain;
-        this.audience = audience;
-        this.audienceId = audienceId;
-    }
-
     /**
      * Retrieve entity type
      *
-     * @return Entity type
+     * @return Entity Type
      */
     public String getEntityType() {
 
@@ -108,63 +94,4 @@ public class Entity {
         this.tenantId = tenantId;
     }
 
-    /**
-     * Get tenant domain of the entity
-     *
-     * @return Tenant domain
-     */
-    public String getTenantDomain() {
-
-        return tenantDomain;
-    }
-
-    /**
-     * Set tenant domain for entity
-     *
-     * @param tenantDomain value to set as tenant domain
-     */
-    public void setTenantDomain(String tenantDomain) {
-
-        this.tenantDomain = tenantDomain;
-    }
-
-    /**
-     * Get audience of the entity
-     *
-     * @return Audience
-     */
-    public String getAudience() {
-
-        return audience;
-    }
-
-    /**
-     * Set audience for entity
-     *
-     * @param audience value to set as audience
-     */
-    public void setAudience(String audience) {
-
-        this.audience = audience;
-    }
-
-    /**
-     * Get audience ID of the entity
-     *
-     * @return Audience ID
-     */
-    public String getAudienceId() {
-
-        return audienceId;
-    }
-
-    /**
-     * Set audience ID for entity
-     *
-     * @param audienceId value to set as audience ID
-     */
-    public void setAudienceId(String audienceId) {
-
-        this.audienceId = audienceId;
-    }
 }
