+++ components/workflow-mgt/org.wso2.carbon.identity.workflow.mgt/src/main/java/org/wso2/carbon/identity/workflow/mgt/util/WorkflowErrorConstants.java
@@ -41,18 +41,16 @@ public class WorkflowErrorConstants {
                 "The user already exists in the system."),
         ERROR_CODE_USER_WF_ROLE_NOT_FOUND("WFM-10006",
                 "The role %s is not found in the system for assign the user."),
-        ERROR_CODE_USER_WF_ROLE_PENDING_APPROVAL("WFM-10007",
-                "There is the pending deletion/update workflow for the role: %s"),
+        ERROR_CODE_USER_WF_ROLE_PENDING_DELETION("WFM-10007",
+                "There is the pending deletion workflow for the role: %s"),
 
         ERROR_CODE_ROLE_WF_ALREADY_EXISTS("WFM-10008",
                 "There is a pending workflow already defined for the role."),
         ERROR_CODE_ROLE_WF_ROLE_ALREADY_EXISTS("WFM-10009",
                 "The role already exist in the system."),
-        ERROR_CODE_ROLE_WF_ROLE_NOT_FOUND("WFM-10010",
-                "The role is not found in the system."),
-        ERROR_CODE_ROLE_WF_USER_NOT_FOUND("WFM-10011",
+        ERROR_CODE_ROLE_WF_USER_NOT_FOUND("WFM-10010",
                 "The user %s is not found in the system for assign the role."),
-        ERROR_CODE_ROLE_WF_USER_PENDING_APPROVAL("WFM-10012",
+        ERROR_CODE_ROLE_WF_USER_PENDING_DELETION("WFM-10011",
                 "There is the pending deletion workflow for the user: %s");
         private final String code;
         private final String message;
